Add check for libtirpc in place of the old glibc rpc calls
I've added an external dependency for tirpc since glibc got rid of the
old Sun RPC calls quite some time ago (though this code is contemporary
to their existence in glibc). An include path has also been added here,
but it would be wise to refactor it to dynamically find the tirpc
headers sometime in the future. For now, this should cover most cases.

Signed-off-by: devfonks <noah@nkeck72.xyz>